% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_table_freq.R
\name{add_table_freq}
\alias{add_table_freq}
\title{Make frequencies table}
\usage{
add_table_freq(lsci)
}
\arguments{
\item{lsci}{An object of class scimeetr}
}
\value{
An object of class scimeetr
}
\description{
Make frequencies table
}
