% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.scimeetr.R
\name{as.scimeetr}
\alias{as.scimeetr}
\title{From data frame to scimeetr}
\usage{
as.scimeetr(dfsci)
}
\arguments{
\item{dfsci}{a dataframe derived from scimeetr}
}
\value{
an object of class scimeetr.
}
\description{
From data frame to scimeetr
}
