% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterize_ab.R
\name{characterize_ab}
\alias{characterize_ab}
\title{Make a list of dataframe of abstract word frequency}
\usage{
characterize_ab(scimeetr_data, lambda = 0.4)
}
\arguments{
\item{scimeetr_data}{A scimeetr object}
}
\value{
A list of dataframe of length equal to the number of sub-communities
}
\description{
This can be used to characterize the sub-communities found by scimap().
}
