% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterize_au.R
\name{characterize_au}
\alias{characterize_au}
\title{Make a list of dataframe of Authors frequencies}
\usage{
characterize_au(scimeetr_data)
}
\arguments{
\item{scimeetr_data}{A scimeetr object}
}
\value{
A list of dataframe containing Author related metrics
}
\description{
This can be used to characterize the communities of research.
}
