% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterize_co.R
\name{characterize_co}
\alias{characterize_co}
\title{Make a list of dataframe of country frequency}
\usage{
characterize_co(scimeetr_data)
}
\arguments{
\item{scimeetr_data}{A scimeetr object}
}
\value{
A list of dataframe containing journal related metrics
}
\description{
This can be used to characterize the communities of research.
}
