% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterize_kw.R
\name{characterize_kw}
\alias{characterize_kw}
\title{Make a list of dataframe of keyword frequency}
\usage{
characterize_kw(scimeetr_data)
}
\arguments{
\item{scimeetr_data}{A scimeetr object}
}
\value{
A plot, generated from the sankeyNetwork function in the package networkD3
}
\description{
This can be used to characterize the sub-communities found by scimap().
}
