% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterize_ti.R
\name{characterize_ti}
\alias{characterize_ti}
\title{Make a list of dataframe of title word frequency}
\usage{
characterize_ti(scimeetr_data, n = 10)
}
\arguments{
\item{scimeetr_data}{A scimeetr object}
}
\value{
A list of dataframe of length equal to the number of sub-communities
}
\description{
This can be used to characterize the sub-communities found by scimap().
}
