% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterize_un.R
\name{characterize_un}
\alias{characterize_un}
\title{Make a list of dataframe of univeristy frequency}
\usage{
characterize_un(scimeetr_data)
}
\arguments{
\item{scimeetr_data}{A scimeetr object}
}
\value{
A list of dataframe containing university related metrics
}
\description{
This can be used to characterize the communities of research.
}
