% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterize.R
\name{clusterize}
\alias{clusterize}
\title{Wrapper of igraph community algorithms}
\usage{
clusterize(graph = graph, community_algorithm = "louvain")
}
\arguments{
\item{graph}{A igraph object.}

\item{community_algorithm}{A character vector of length one.}
}
\value{
An object of class 'community'.
}
\description{
A function that helps you pick some community clustering algorithms provided
by the package \link{igraph}.
}
