% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_scimap.R
\name{compare_scimap}
\alias{compare_scimap}
\title{Compare the communities in a scimeetr object}
\usage{
compare_scimap(sl)
}
\arguments{
\item{sl}{A list of scimeet object}
}
\value{
A plot, generated from the sankeyNetwork function in the package networkD3
}
\description{
Compare the communities in a scimeetr object
}
