% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupling.R
\name{coupling}
\alias{coupling}
\title{Make graph based on coupling}
\usage{
coupling(dfsci = dfsci, coupling_by = "bic")
}
\arguments{
\item{dfsci}{A data.frame of class scimeetr created from importing files
with import_bib_file or a list of class scimeetr created from the
function scimap.}

\item{coupling_by}{A vector of length one. Equal to either:
'bc'/'bibliographic coupling', 'kc'/'keyword coupling', 'tc'/'title coupling',
'ac'/'abstract coupling', 'jc'/'journal coupling', 'ac'/'author coupling'.}
}
\value{
graph object
}
\description{
Make graph based on coupling
}
