% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_citer.R
\name{find_citer}
\alias{find_citer}
\title{Find publication citer}
\usage{
find_citer(lsci, publication)
}
\arguments{
\item{lsci}{a scimeetr object}

\item{publication}{a vector of length one matchin the name of a publication from scilist output}
}
\value{
A data.frame
}
\description{
Find publication citer
}
