% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_citers_of_other_com.R
\name{find_citers_of_other_com}
\alias{find_citers_of_other_com}
\title{Find papers of one community that cite most the papers in an other community}
\usage{
find_citers_of_other_com(scimeetr_data, com_citing, com_cited)
}
\arguments{
\item{scimeetr_data}{A scimeetr object.}

\item{com_citing}{A character string matching the name of the citing
community.}

\item{com_cited}{A character string matching the name of the cited community.}
}
\value{
A dataframe.
}
\description{
With \code{merge_scopus_and_wos} you can merge bibliometric data from Scopus
and the Web of Science. Be carefull if you are using this prior to
bibliographic coupling. Scopus and WOS format the cited reference field
differently.
}
\details{
I thank Dr. Sofia van Moorsel for stimulating the creation of this
  function.
}
