% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focus_on.R
\name{focus_on}
\alias{focus_on}
\title{Focus on only one community}
\usage{
focus_on(scimeetr_list, grab)
}
\arguments{
\item{scimeetr_list}{A \link{scimeetr} object.}

\item{grab}{A character string matching the name of the community that you
which to focus on.}
}
\value{
A scimeetr object.
}
\description{
Focus on only one community
}
