% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_synonyms.R
\name{get_synonyms}
\alias{get_synonyms}
\title{Find similar words}
\usage{
get_synonyms(scimeetr_data, synonymous_to, n_gram = c(1L, 3L))
}
\arguments{
\item{scimeetr_data}{a scimeetr object}

\item{synonymous_to}{an character string that is an expression for which 
synonymous and similar expression are wanted. synonyms for each elements in
the vectors will be proposed.}
}
\value{
a list of character string. In other words, a vector of suggestions.
}
\description{
Find similar words
}
