% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_scopus_files.R
\name{import_scopus_files}
\alias{import_scopus_files}
\title{Import and Merge Scopus Files into a scimeetr list}
\usage{
import_scopus_files(files_directory)
}
\arguments{
\item{files_directory}{a character vector giving the \bold{folder} path in 
which all the Web of Science files to be imported into a data frame can be 
found. This folder should contain \bold{only} the files to be imported.}
}
\value{
data frame of 62 columns and a number of row equal to the number of 
  unique records.
}
\description{
This function imports and if there is more than one file in the user selected
directory automatically merges Web of Science files. This is all put into a 
data frame. Moreover, all duplicated records are automatically removed.
}
\details{
No details for now.
}
\seealso{
\code{\link{scimeetr}} and \code{\link{import_wos_files}}.
}
\author{
Maxime Rivest
}
\keyword{manip}
