% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_recid.R
\name{make_recid}
\alias{make_recid}
\title{Make a recid that math the RECID from the function split_cr()}
\usage{
make_recid(WOS_table)
}
\arguments{
\item{dfsci}{a dataframe with a CR column}
}
\value{
a vector of same length as input df
}
\description{
Make a recid that math the RECID from the function split_cr()
}
