% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_scimaps.R
\name{merge_scimaps}
\alias{merge_scimaps}
\title{Merge scimaps}
\usage{
merge_scimaps(sl, minimum_com_overlap, min_com_size = 15)
}
\arguments{
\item{sl}{A list of \link{scimeetr} object}

\item{minimum_com_overlap}{An integer. Minimum weight of the link between 
paper to keep in the graph. If it is equal to length(sl) that means that 
papers that were not classifies in the same communities for every scimap 
would not be linked in a network.}

\item{min_com_size}{An integer.}
}
\value{
A list a class scimeetr.
}
\description{
With \code{merge_scimaps} you can re-assign communities to papers based on a 
conscensus from different coupling options.
}
