% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_scopus_and_wos.R
\name{merge_scopus_and_wos}
\alias{merge_scopus_and_wos}
\title{Merge scimaps}
\usage{
merge_scopus_and_wos(scopus, wos)
}
\arguments{
\item{scopus}{A \link{scimeetr} object from \link{import_scopus_files}()}

\item{wos}{A \link{scimeetr} object from \link{import_wos_files}()}
}
\value{
A list of class scimeetr.
}
\description{
With \code{merge_scimaps} you can re-assign communities to papers based on a 
conscensus from different coupling options.
}
