% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.scimeetrsum.R
\name{plot.scimeetrsum}
\alias{plot.scimeetrsum}
\title{Plot scimeetr summary}
\usage{
\method{plot}{scimeetrsum}(object, node_names = "tag", ...)
}
\arguments{
\item{object}{}

\item{...}{}
}
\description{
Plot scimeetr summary
}
