% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_publication_network.R
\name{plot_publication_network}
\alias{plot_publication_network}
\title{Plot publication network colored by sub-communities}
\usage{
plot_publication_network(scimeetr_data, plot_to_file = F, filename, width,
  height, min_com_size = 100)
}
\arguments{
\item{scimeetr_data}{A \link{scimeetr} object}

\item{plot_to_file}{True or False}

\item{filename}{a character string matchin a path name}

\item{width}{Argument to jpeg()}

\item{height}{Argument to jpeg()}

\item{min_com_size}{A numeric vector of length one.}

\item{...}{if plot_to_file is True, parameter of the jpeg() function.}
}
\value{
A plot, generated from the sankeyNetwork function in the package
  networkD3
}
\description{
With \code{plot_publication_network} you can generate a plot of a publication
network which will be colored by sub-communities.
}
