% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_subcommunities.R
\name{plot_subcommunities}
\alias{plot_subcommunities}
\title{Plot network of sub-communities}
\usage{
plot_subcommunities(scimeetr, scimeetr_id = "com1")
}
\arguments{
\item{scimeetr}{a scimeetr object}

\item{scimeetr_id}{a community id that contains a graph}
}
\value{
a Networkvis plot
}
\description{
Plot network of sub-communities
}
