% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.scimeetrsum.R
\name{print.scimeetrsum}
\alias{print.scimeetrsum}
\title{Print scimeetr summary}
\usage{
\method{print}{scimeetrsum}(object, ...)
}
\arguments{
\item{object}{}

\item{...}{}
}
\description{
Print scimeetr summary
}
