% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl2dfsci.R
\name{rl2dfsci}
\alias{rl2dfsci}
\title{Find list content in dfsci database}
\usage{
rl2dfsci(lsci, rl)
}
\arguments{
\item{lsci}{a scimeetr object}

\item{rl}{a reading list from scilist}
}
\value{
A data.frame
}
\description{
Find list content in dfsci database
}
