% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scilist_all.R
\name{scilist_all}
\alias{scilist_all}
\title{Make all reading lists}
\usage{
scilist_all(scimeetr_data, length_list = 1, verbose = T,
  except_bet_clo = T)
}
\arguments{
\item{scimeetr_data}{A scimeetr object. This is created from importing files
with import_wos_file or scimap().}

\item{length_list}{An integer that will multiply the standard list length.}
}
\value{
a list of reading list of the same length as the number of
  communities found in the input scimeetr object
}
\description{
Make all reading lists
}
