% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scimap.R
\name{scimap}
\alias{scimap}
\title{Finds research communities}
\usage{
scimap(scimeetr_list = scimeetr_list, coupling_by = "bic",
  community_algorithm = "louvain", min_com_size = 30)
}
\arguments{
\item{scimeetr_list}{A list of class scimeetr. Objects of class scimeetr are return by 
with import_bib_file or a list of class scimeetr created from the 
function scimap.}

\item{community_algorithm}{A vector of length one. 'louvain', 'fast greedy'...}

\item{coupling}{A vector of length one. Equal to either: 
'bc'/'bibliographic coupling', 'kc'/'keyword coupling', 'tc'/'title coupling', 
'ac'/'abstract coupling', 'jc'/'journal coupling', 'ac'/'author coupling'.}

\item{inclusive_com}{TRUE or FALSE. Tells wheter to include within the subcommunities
all the frontier nodes that were not originally included in the subcommunities.}
}
\value{
biblio_list. A list of the communities, including the global community,
 with all their associated graph object, community object and biblio_df object.
}
\description{
Finds research communities
}
