% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scishine.R
\name{scishine}
\alias{scishine}
\title{Start shiny app to use scimeetr interactively}
\usage{
scishine(files_directory)
}
\value{
A shiny app
}
\description{
This function starts an interactive web app
}
\details{
No details for now.
}
\examples{
scishine()
}
\seealso{
\code{\link{scimeetr}}.
}
\author{
Maxime Rivest
}
