% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_cr.R
\name{split_cr}
\alias{split_cr}
\title{Split cited references into first author, year, journal, etc.}
\usage{
split_cr(lsci, min_cr_freq = 2)
}
\arguments{
\item{dfsci}{a dataframe with a CR column}
}
\value{
a dataframe
}
\description{
Split cited references into first author, year, journal, etc.
}
