% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.scimeetr.R
\name{summary.scimeetr}
\alias{summary.scimeetr}
\title{Summary scimeetr}
\usage{
\method{summary}{scimeetr}(object, com_size = 200, ...)
}
\arguments{
\item{object}{a scimeetr object}

\item{com_size}{minimum size of a community to plot}

\item{...}{options to plot}
}
\value{
object of class scimeetrSum
}
\description{
Summary scimeetr
}
