\name{wos_df_example}
\alias{wos_df_example}
\docType{data}
\title{
Web of Science dataframe
}
\description{
to come \dots
}
\usage{data("wos_df_example")}
\format{
  A data frame with 50 observations on the following 62 variables.
  \describe{
    \item{\code{PT}}{a character vector}
    \item{\code{AU}}{a character vector}
    \item{\code{BA}}{a logical vector}
    \item{\code{BE}}{a character vector}
    \item{\code{GP}}{a logical vector}
    \item{\code{AF}}{a character vector}
    \item{\code{BF}}{a logical vector}
    \item{\code{CA}}{a logical vector}
    \item{\code{TI}}{a character vector}
    \item{\code{SO}}{a character vector}
    \item{\code{SE}}{a character vector}
    \item{\code{BS}}{a logical vector}
    \item{\code{LA}}{a character vector}
    \item{\code{DT}}{a character vector}
    \item{\code{CT}}{a character vector}
    \item{\code{CY}}{a character vector}
    \item{\code{CL}}{a character vector}
    \item{\code{SP}}{a character vector}
    \item{\code{HO}}{a logical vector}
    \item{\code{DE}}{a character vector}
    \item{\code{ID}}{a character vector}
    \item{\code{AB}}{a character vector}
    \item{\code{C1}}{a character vector}
    \item{\code{RP}}{a character vector}
    \item{\code{EM}}{a character vector}
    \item{\code{RI}}{a character vector}
    \item{\code{OI}}{a character vector}
    \item{\code{FU}}{a character vector}
    \item{\code{FX}}{a character vector}
    \item{\code{CR}}{a character vector}
    \item{\code{NR}}{a numeric vector}
    \item{\code{TC}}{a numeric vector}
    \item{\code{Z9}}{a numeric vector}
    \item{\code{U1}}{a numeric vector}
    \item{\code{U2}}{a numeric vector}
    \item{\code{PU}}{a character vector}
    \item{\code{PI}}{a character vector}
    \item{\code{PA}}{a character vector}
    \item{\code{SN}}{a character vector}
    \item{\code{EI}}{a character vector}
    \item{\code{BN}}{a logical vector}
    \item{\code{J9}}{a character vector}
    \item{\code{JI}}{a character vector}
    \item{\code{PD}}{a character vector}
    \item{\code{PY}}{a numeric vector}
    \item{\code{VL}}{a numeric vector}
    \item{\code{IS}}{a character vector}
    \item{\code{PN}}{a logical vector}
    \item{\code{SU}}{a logical vector}
    \item{\code{SI}}{a logical vector}
    \item{\code{MA}}{a logical vector}
    \item{\code{BP}}{a numeric vector}
    \item{\code{EP}}{a numeric vector}
    \item{\code{AR}}{a character vector}
    \item{\code{DI}}{a character vector}
    \item{\code{D2}}{a logical vector}
    \item{\code{PG}}{a numeric vector}
    \item{\code{WC}}{a character vector}
    \item{\code{SC}}{a character vector}
    \item{\code{GA}}{a character vector}
    \item{\code{UT}}{a character vector}
    \item{\code{PM}}{a numeric vector}
  }
}
\details{
none
}
\source{
Web of Science
}
\references{
none
}
\examples{
data(wos_df_example)
}
\keyword{datasets}
