% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterize_au.R
\name{characterize_au}
\alias{characterize_au}
\title{Characterize authors found within papers' titles}
\usage{
characterize_au(scimeetr_data, lambda = 0.7)
}
\arguments{
\item{scimeetr_data}{An object of class scimeetr.}

\item{lambda}{A number from 0 to 1. 0 for relative frequency 1 for total 
occurence only}
}
\value{
A list of dataframe. The list length matchs the number of communities
  that the scimeetr object contains.
}
\description{
\code{characterize_au} calculates several author related metrics from a
scimeetr object. The results are returned in a list of data frame. The
metrics in the table are: title-words frequency, title-words relative
frequency, title-words relevance.
}
\examples{
# Example with an object of class scimeetr (see import_wos_files() or 
# import_scopus_files()) already in the workspace
author_list <- characterize_au(scimeetr_list)
# Since this example shows how to load WOS from your system we need to run 
# the following line to find the path to the exemple file
fpath <- system.file("extdata", package="scimeetr") 
fpath <- paste(fpath, "/wos_folder/", sep = "") 
# Then we can run the actual example
example_scimeetr_object <- import_wos_files(files_directory = fpath)
characterize_au(example_scimeetr_object)

}
\seealso{
\code{\link{characterize_jo}} for journal characterization, 
  \code{\link{characterize_ab}} for abstract-word characterization, 
  \code{\link{characterize_kw}} for keyword characterization, 
  \code{\link{characterize_ti}} for title characterization, 
  \code{\link{characterize_un}} for university characterization, 
  \code{\link{characterize_co}} for country characterization
}
