% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dive_to.R
\name{dive_to}
\alias{dive_to}
\title{Dive to a sub-community and keep its sub-communities}
\usage{
dive_to(scimeetr_list, aim_at)
}
\arguments{
\item{scimeetr_list}{A \link{scimeetr} object}

\item{aim_at}{A character string matching the name of the community that you
which to get to}
}
\value{
A scimeetr object.
}
\description{
Dive to a sub-community and keep its sub-communities
}
