% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_scopus_and_wos.R
\name{merge_scopus_and_wos}
\alias{merge_scopus_and_wos}
\title{Merge data from scopus and Web of Science together}
\usage{
merge_scopus_and_wos(scopus, wos)
}
\arguments{
\item{scopus}{A \link{scimeetr} object from \link{import_scopus_files}()}

\item{wos}{A \link{scimeetr} object from \link{import_wos_files}()}
}
\value{
A list of class scimeetr.
}
\description{
With \code{merge_scopus_and_wos} you can merge bibliometric data from Scopus
and the Web of Science. Be carefull if you are using this prior to
bibliographic coupling. Scopus and WOS format the cited reference field
differently.
}
\details{
I thank Dr. Shinichi Nakagawa and Rose O'Dea for requiring this
  feature.
}
