% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scimap.R
\name{scimap}
\alias{scimap}
\title{Finds research communities}
\usage{
scimap(scimeetr_data, coupling_by = "bic", community_algorithm = "louvain",
  min_com_size = 30, kw = 1, ti = 1, ab = 1)
}
\arguments{
\item{scimeetr_data}{A list of class scimeetr. Objects of class scimeetr are 
returned by import_wos_files or by scimap itself.}

\item{coupling_by}{A character vector of length one. Equal to either: 'bic' 
for bibliographic coupling, 'kec' for keyword coupling, 'tic' for title 
coupling, 'abc' for abstract coupling, 'joc' for journal coupling, 'auc' 
for author coupling', 'woc' for word coupling, 'bickec' for a combination
of bibliographic coupling and keyword coupling, and 'bickecticjoc' for a
combination of bic, kec, tic, joc (this might be the better way for most
use but it is a bit slow, this is why it is not the default choice).}

\item{community_algorithm}{A character vector of length one. 'louvain', 'fast
greedy'...}
}
\value{
scimeetr_data. A \link{scimeetr} object.
}
\description{
\code{scimap} is used to find sub-communities of a \link{scimeetr} object.
}
\examples{
# Here we see that scimeetr_list contains no sub-communities
summary(scimeetr_list)
sci <- scimap(scimeetr_list, coupling_by = 'bickec')
# After applying scimap() to scimeetr_list. The resulting variable (sci)
# contains sub-communities
summary(sci)
}
