% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_cr.R
\name{split_cr}
\alias{split_cr}
\title{Split cited references into first author, year, journal, etc.}
\usage{
split_cr(scimeetr_data, min_cr_freq = 2)
}
\arguments{
\item{scimeetr_data}{A scimeetr object.}
}
\value{
a dataframe
}
\description{
Split cited references into first author, year, journal, etc.
}
